+++ src/data/memorialPlacesData.ts
@@ -4,26 +4,11 @@ import type { PlaceDetails } from "@/types";
 
 export const memorialPlaces: PlaceDetails[] = [
   {
-    id: 1,
-    name: "Metalowy krzyż upamiętniający żołnierzy 5. Pułku Strzelców Konnych",
-    location: "Cmentarz Parafialny w Dębicy",
-    description:
-      "Krzyż upamiętniający zakończenie II wojny światowej oraz rozwiązanie Armii Krajowej. Poświęcony w 1975 roku, z okazji 30. rocznicy tych wydarzeń.",
-    highlight: [
-      { text: "Pierwszy paragraf tekstu w zielonym polu", align: "center" },
-      { text: "Drugi paragraf", align: "left" },
-      { text: "Trzeci paragraf", align: "right" },
-    ],
-    descriptionContinuation: "Dalszy ciąg opisu miejsca.",
-    image: "/images/place1.jpg",
-    link: "/zwiazek/miejsca-pamieci/1",
-  },
-  {
-    id: 2,
+    slug: "pomnik-pchor.-mariana-chmielewskiego",
     name: "Pomnik upamiętniający pchor. Mariana Chmielewskiego ps. „Maria”",
     location: "Dębica",
     description:
-      "Pomnik pchor. Mariana Chmielewskiego został poświęcony 18 sierpnia 1996 roku. Inicjatorami upamiętnienia byli koledzy z plutonu dywersyjnego i członkowie Światowego Związku Żołnierzy Armii Krajowej w Dębicy. Pomnik uległ zniszczeniu, a tablica upamiętniająca bohatera, prawdopodobnie podczas prac budowlanych na stacji kolejowej, zaginęła. Z tej racji Zarząd ŚZŻAK Środowiska 5. PSK AK w Dębicy w lipcu 2014 roku podjął decyzję o odnowieniu pomnika i ufundowaniu nowej tablicy.      W pomoc włączyła się gmina Czarna. Uroczystość odsłonięcia odnowionego obelisku odbyła się 17 sierpnia 2014 roku i rozpoczęła się mszą świętą sprawowaną przy grocie NMP w Straszęcinie.Następnie uczestnicy uroczystości przejechali pod stację kolejową w Grabinach gdzie odbyła się część oficjalna uroczystości. Odnowiony pomnik poświęcił ks. Bogusław Czech, a odsłonięcia monumentu dokonali: por. Izabela Podhalańska i wójt gminy Józef Chudy. Następnie głos zabrał Maciej Małozięć, wiceprezes Okręgu Tarnów i inicjator odnowienia pomnika, który przypomniał postać pchor. Mariana Chmielewskiego ps. „Maria”. Na tablicy znajduje się napis: ",
+      "Pomnik pchor. Mariana Chmielewskiego został poświęcony 18 sierpnia 1996 roku. Inicjatorami upamiętnienia byli koledzy z plutonu dywersyjnego i członkowie Światowego Związku Żołnierzy Armii Krajowej w Dębicy. Pomnik uległ zniszczeniu, a tablica upamiętniająca bohatera, prawdopodobnie podczas prac budowlanych na stacji kolejowej, zaginęła. Z tej racji Zarząd ŚZŻAK Środowiska 5. PSK AK w Dębicy w lipcu 2014 roku podjął decyzję o odnowieniu pomnika i ufundowaniu nowej tablicy. W pomoc włączyła się gmina Czarna. Uroczystość odsłonięcia odnowionego obelisku odbyła się 17 sierpnia 2014 roku i rozpoczęła się mszą świętą sprawowaną przy grocie NMP w Straszęcinie.Następnie uczestnicy uroczystości przejechali pod stację kolejową w Grabinach gdzie odbyła się część oficjalna uroczystości. Odnowiony pomnik poświęcił ks. Bogusław Czech, a odsłonięcia monumentu dokonali: por. Izabela Podhalańska i wójt gminy Józef Chudy. Następnie głos zabrał Maciej Małozięć, wiceprezes Okręgu Tarnów i inicjator odnowienia pomnika, który przypomniał postać pchor. Mariana Chmielewskiego ps. „Maria”. Na tablicy znajduje się napis: ",
     highlight: [
       {
         text: "PAMIĘCI\nPCHOR. MARIANA CHMIELEWSKIEGO\nPS. „MARIA” ŻOŁNIERZA ARMII KRAJOWEJ\nPOLEGŁEGO W WALCE Z NIEMIECKĄ\nPOLICJĄ KOLEJOWĄ W DNIU 19 VIII 1943 R.",
@@ -40,66 +25,20 @@ export const memorialPlaces: PlaceDetails[] = [
     ],
     descriptionContinuation: " ",
     image: "/images/memorial-places/place2.jpg",
-    link: "/zwiazek/miejsca-pamieci/2",
   },
-  {
-    id: 3,
-    name: "Miejsce 3",
-    location: "Lokalizacja miejsca 3",
-    description:
-      "Lorem ipsum dolor sit amet consectetur. Vitae ornare odio et ullamcorper arcu cursus.",
-    highlight: [
-      { text: "Pierwszy paragraf tekstu w zielonym polu", align: "center" },
-      { text: "Drugi paragraf", align: "left" },
-      { text: "Trzeci paragraf", align: "right" },
-    ],
-    descriptionContinuation: "Dalszy ciąg opisu miejsca.",
-    image: "/images/place3.jpg",
-    link: "/zwiazek/miejsca-pamieci/3",
-  },
-  {
-    id: 4,
-    name: "Miejsce 4",
-    location: "Lokalizacja miejsca 4",
-    description:
-      "Lorem ipsum dolor sit amet consectetur. Vitae ornare odio et ullamcorper arcu cursus.",
-    highlight: [
-      { text: "Pierwszy paragraf tekstu w zielonym polu", align: "center" },
-      { text: "Drugi paragraf", align: "left" },
-      { text: "Trzeci paragraf", align: "right" },
-    ],
-    descriptionContinuation: "Dalszy ciąg opisu miejsca.",
-    image: "/images/place4.jpg",
-    link: "/zwiazek/miejsca-pamieci/4",
-  },
-  {
-    id: 5,
-    name: "Miejsce 5",
-    location: "Lokalizacja miejsca 5",
-    description:
-      "Lorem ipsum dolor sit amet consectetur. Vitae ornare odio et ullamcorper arcu cursus.",
-    highlight: [
-      { text: "Pierwszy paragraf tekstu w zielonym polu", align: "center" },
-      { text: "Drugi paragraf", align: "left" },
-      { text: "Trzeci paragraf", align: "right" },
-    ],
-    descriptionContinuation: "Dalszy ciąg opisu miejsca.",
-    image: "/images/place5.jpg",
-    link: "/zwiazek/miejsca-pamieci/5",
-  },
-  {
-    id: 6,
-    name: "Miejsce 6",
-    location: "Lokalizacja miejsca 6",
-    description:
-      "Lorem ipsum dolor sit amet consectetur. Vitae ornare odio et ullamcorper arcu cursus.",
-    highlight: [
-      { text: "Pierwszy paragraf tekstu w zielonym polu", align: "center" },
-      { text: "Drugi paragraf", align: "left" },
-      { text: "Trzeci paragraf", align: "right" },
-    ],
-    descriptionContinuation: "Dalszy ciąg opisu miejsca.",
-    image: "/images/place6.jpg",
-    link: "/zwiazek/miejsca-pamieci/6",
-  },
+  // {
+  //   slug: "",
+  //   name: "Miejsce 3",
+  //   location: "Lokalizacja miejsca 3",
+  //   description:
+  //     "Lorem ipsum dolor sit amet consectetur. Vitae ornare odio et ullamcorper arcu cursus.",
+  //   highlight: [
+  //     { text: "Pierwszy paragraf tekstu w zielonym polu", align: "center" },
+  //     { text: "Drugi paragraf", align: "left" },
+  //     { text: "Trzeci paragraf", align: "right" },
+  //   ],
+  //   descriptionContinuation: "Dalszy ciąg opisu miejsca.",
+  //   image: "/images/place3.jpg",
+  //   link: "/zwiazek/miejsca-pamieci/${slug}",
+  // },
 ];
